支持小数边权（解析与 Prim 求解器）

+++ include/graph.h
@@ -7,7 +7,7 @@ struct Edge {
     int id;
     int u;
     int v;
-    int weight;
+    double weight;
 };
 
 class IGraphStorage {
@@ -18,25 +18,25 @@ public:
     virtual const QVector<Edge>& edges() const = 0;
     virtual QVector<int> neighbors(int vertex) const = 0;
     virtual int edgeId(int u, int v) const = 0;
-    virtual int edgeWeight(int u, int v) const = 0;
+    virtual double edgeWeight(int u, int v) const = 0;
 };
 
 class AdjMatrixGraph final : public IGraphStorage {
 public:
     explicit AdjMatrixGraph(int n = 0);
     void reset(int n);
-    void addEdge(int u, int v, int w);
+    void addEdge(int u, int v, double w);
 
     int vertexCount() const override;
     int edgeCount() const override;
     const QVector<Edge>& edges() const override;
     QVector<int> neighbors(int vertex) const override;
     int edgeId(int u, int v) const override;
-    int edgeWeight(int u, int v) const override;
+    double edgeWeight(int u, int v) const override;
 
 private:
     int m_n;
-    QVector<QVector<int>> m_weight;
+    QVector<QVector<double>> m_weight;
     QVector<QVector<int>> m_edgeId;
     QVector<Edge> m_edges;
 };
@@ -45,19 +45,19 @@ class AdjListGraph final : public IGraphStorage {
 public:
     explicit AdjListGraph(int n = 0);
     void reset(int n);
-    void addEdge(int u, int v, int w);
+    void addEdge(int u, int v, double w);
 
     int vertexCount() const override;
     int edgeCount() const override;
     const QVector<Edge>& edges() const override;
     QVector<int> neighbors(int vertex) const override;
     int edgeId(int u, int v) const override;
-    int edgeWeight(int u, int v) const override;
+    double edgeWeight(int u, int v) const override;
 
 private:
     struct To {
         int v;
-        int w;
+        double w;
         int id;
     };
